requirements.txt
paddleocr
paddlepaddle==3.3.1
opencv-python
pydantic
fastapi
uvicorn
python-multipart
Pillow
